Updated color scheme and fixed bug in e/pi buttons
Updated the buttons in portrait and landscape mode to have a yellow color. Also updated the background to be a dark gray. Updated text to be white, which is easier to read on a dark background. I also noticed the actions for the "e" button and the "pi" button were swapped, so I swapped them back