Updated updateText function
TL;DR: The cursor in the equation text now always stays to the far right.

Before, when strings that were longer than 1 character were added to the textview, the cursor would only move 1 character to the right, resulting in it still being in the middle of the strings. I updated it so that the updateText function also looks at the length of the string it's adding to the equation, and will move the cursor to the right that many characters.